Move to private, store permanent values in final statics.
